Copy constructor for TexData
More adjustments to texturing, if allowances are made for exact border values, zeros pop into the factors.  So remove the zeros

FreeAll() added to a bunch of stuff

Better smoother threading in Terrain

Print coordinates and thread status in the render window

Use the existing tex data whenever rebuilding a new terrain

save transition height

Save all grid data to a file and reload it automagically.  Should be ok if textures change and such.