Got rid of a lot more cruft that was mainly needed for that failed attempt at removing duplicate verts
Replaced a lot of asserts with printed warnings to the new sharedforms output window stolen out of the bsp builder project.  This was really needed as there are a lot of small cases that the artist needs to know about but aren't necessarily fatal.  Also the messages can explain what happened when things outright fail.  Before nothing would happen and you had no idea why.

Still problems with animations to fix but getting closer.