Moved BSPBuilder's debug draw module out to SharedForms so the light explorer can use it
Got the light recording stuff working again that records extra data and saves it out for the light explorer to load up

Resurrection of LightExplorer, a tool to view the per-face details of the lighting process.  This was originally an XNA project so little of it could still be used.  PGUP/Down (shift for 100 at a time) to skip faces, X to hide world geometry (drawn via debugdraw).