Seeing weights that don't add up to 1 coming from blender.   Probably not a good idea but I decided to do a pass on the weights to make them add up to 1.  Nuked the old way of adding in the contribution of the deleted weights in case of the number of influences being greater than 4.
The polygon indexing stuff in AddNormTexByPoly() was kind of pointless, as was the index list in there as verts are already duped by poly.  Lazy but it works.

Triangulation now works.  Blender outputs quads.  It also does indexes into lists differently.  It seems to interpret the spec a different way than max/maya.  I have it working but I fear I broke max/maya.  (This stuff isn't in yet)