Using exports with z forward and Y up, I finally managed to get rid of all the strange 90 degree rotations, vert axis flipping, and superweird matrix column negation stuff I was doing.
Most of the shiftmat stuff is gone, as is the adjustment of the skeleton and inverse bind poses.  All that is there is the scale for unit conversion.

Blender's collada exporter doesn't obey the format's specifier for axis stuff, but it doesn't really matter as long as the export options are correct.