Lots of changes to make things work after ditching the arch interface and all that mess.
Still lots broken but you can load a DAE and animate and such

Lots of ongoing work to get rid of warnings.  I've always assumed any ref type can be null, but now the language wants you to do ?vars if you need that sort of behaviour.  It's a pain but I guess it is more readable eventually, and maybe safe?  Long way to go to fix it all though.