Static normals needed an axis swap instead of just a rotate.  Different handedness.
Switching arrrow keys to be used as a way to move collision nodes in edit mode

Edit mode for collision nodes, similar to the way the bone collision shapes work on a character.  Not too sure it is the right way to do it though.  Another way I'm considering would be to click-pick a set of faces with the mouse, then hit a key to bound those and make a node out of it.