Added a righthand switch to ccnogui to save stuff in right handed format.  Otherwise it saves in Left.
Fix additions to an animlib with the animation already existing.  I
blast the existing anim and add the new one.  If there's only one anim
that will nuke the skeleton as well so I check for that and re-add the
skeleton if needed.

Left and right versions of the root transform func.  Not working yet.
Still need to get things unmirrored and oriented properly.